pyproject.toml
[project]
name = "nonebot_plugin_anan_say"
version = "0.1.0"
description = "一个向安安的素描本上渲染文字并发送出去的插件"
authors = [
    {name = "Mr.Lee", email = "chzxxuanzheng@gmail.com"},
]
dependencies = [
    "nonebot-plugin-alconna>=0.59.4",
    "nonebot2>=2.4.3",
    "pillow>=11.3.0",
]
requires-python = ">=3.9"
readme = "README.md"
license = {text = "agpl"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
